Remove test and merge
